refactor: cdk support
Add documentation and move both CDK utilities into the same cdk utility
file. Add some unit test coverage to verify output is generating.

Fixes #517
